# docs/swagger/swagger.yaml
basePath: /
definitions:
  admin.AdItem:
    properties:
      ad_key:
        type: string
      content_code:
        type: string
      content_url:
        type: string
      duration:
        type: integer
      id:
        type: integer
      link_url:
        type: string
      scene:
        type: string
      sort:
        type: integer
      status:
        type: integer
      title:
        type: string
      type:
        type: string
    type: object
...
  admin.CategoryResponse:
    properties:
      children:
        items:
          $ref: '#/definitions/admin.CategoryResponse'
        type: array
      id:
        type: integer
      name:
        type: string
      parent_id:
        type: integer
      sort_order:
        type: integer
      status:
        type: integer
    type: object
  admin.ChangePasswordRequest:
    properties:
      new_password:
        maxLength: 72
        minLength: 6
        type: string
      old_password:
        maxLength: 72
        minLength: 6
        type: string
    required:
    - new_password
    - old_password
    type: object
...
  admin.CreateAdRequest:
    properties:
      ad_key:
        maxLength: 50
        minLength: 1
        type: string
      content_code:
        maxLength: 10000
        type: string
      content_url:
        maxLength: 500
        type: string
      duration:
        maximum: 300
        minimum: 1
        type: integer
      link_url:
        maxLength: 500
        type: string
      scene:
        enum:
        - video_loading
        - general
        type: string
      sort:
        minimum: 0
        type: integer
      status:
        enum:
        - 0
        - 1
        type: integer
      title:
        maxLength: 128
        minLength: 1
        type: string
      type:
        enum:
        - image
        - video
        - html
        - code
        type: string
    required:
    - ad_key
    - scene
    - title
    - type
    type: object
...
  admin.CreateCategoryRequest:
    properties:
      name:
        maxLength: 100
        minLength: 1
        type: string
      parent_id:
        minimum: 0
        type: integer
      sort_order:
        minimum: 0
        type: integer
      status:
        enum:
        - 0
        - 1
        type: integer
    required:
    - name
    type: object
...
  admin.LoginRequest:
    properties:
      password:
        maxLength: 72
        minLength: 6
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
  admin.LoginResponse:
    properties:
      access_token:
        type: string
      admin:
        $ref: '#/definitions/admin.Profile'
      expires_in:
        type: integer
      token_type:
        type: string
    type: object
...
  admin.Profile:
    properties:
      avatar:
        type: string
      email:
        type: string
      id:
        type: integer
      nickname:
        type: string
      role:
        type: string
      status:
        type: integer
      username:
        type: string
    type: object
...
  admin.UpdateAdRequest:
    properties:
      ad_key:
        maxLength: 50
        minLength: 1
        type: string
      content_code:
        maxLength: 10000
        type: string
      content_url:
        maxLength: 500
        type: string
      duration:
        maximum: 300
        minimum: 1
        type: integer
      link_url:
        maxLength: 500
        type: string
      scene:
        enum:
        - video_loading
        - general
        type: string
      sort:
        minimum: 0
        type: integer
      status:
        enum:
        - 0
        - 1
        type: integer
      title:
        maxLength: 128
        minLength: 1
        type: string
      type:
        enum:
        - image
        - video
        - html
        - code
        type: string
    type: object
...
  admin.UpdateCategoryRequest:
    properties:
      name:
        maxLength: 100
        minLength: 1
        type: string
      parent_id:
        minimum: 0
        type: integer
      sort_order:
        minimum: 0
        type: integer
      status:
        enum:
        - 0
        - 1
        type: integer
    type: object
...
  admin.UpdateProfileRequest:
    properties:
      avatar:
        maxLength: 500
        type: string
      email:
        maxLength: 100
        type: string
      nickname:
        maxLength: 50
        type: string
    type: object
...
  response.PageData:
    properties:
      list: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  response.Response:
    properties:
      cause:
        type: string
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: 小橘TV - 影视站系统
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: 小橘TV
  version: "1.0"
paths:
...
  /api/admin/v1/ads:
    get:
      consumes:
      - application/json
      description: 分页获取广告列表
      parameters:
      - in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - in: query
        maximum: 1000000
        minimum: 1
        name: page
        type: integer
      - in: query
        maximum: 100
        minimum: 1
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PageData'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/admin.AdItem'
                        type: array
                    type: object
              type: object
      security:
      - BearerAuth: []
      summary: 广告列表
      tags:
      - 管理端｜广告管理
    post:
      consumes:
      - application/json
      description: 创建一个新的广告
      parameters:
      - description: 广告参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.CreateAdRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/admin.AdItem'
              type: object
      security:
      - BearerAuth: []
      summary: 新建广告
      tags:
      - 管理端｜广告管理
  /api/admin/v1/ads/{id}:
    delete:
      description: 删除指定广告
      parameters:
      - description: 广告ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除广告
      tags:
      - 管理端｜广告管理
    put:
      consumes:
      - application/json
      description: 更新指定广告信息
      parameters:
      - description: 广告ID
        in: path
        name: id
        required: true
        type: integer
      - description: 广告参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.UpdateAdRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/admin.AdItem'
              type: object
      security:
      - BearerAuth: []
      summary: 更新广告
      tags:
      - 管理端｜广告管理
...
  /api/admin/v1/auth/login:
    post:
      consumes:
      - application/json
      description: 管理员账号密码登录，返回 JWT
      parameters:
      - description: 登录请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/admin.LoginResponse'
              type: object
      summary: 管理员登录
      tags:
      - 管理端｜管理员认证
  /api/admin/v1/auth/logout:
    post:
      consumes:
      - application/json
      description: 无状态登出，客户端清除 token 即可
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 管理员登出
      tags:
      - 管理端｜管理员认证
  /api/admin/v1/auth/profile:
    get:
      description: 获取当前登录管理员的个人资料
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/admin.Profile'
              type: object
      security:
      - BearerAuth: []
      summary: 获取管理员个人资料
      tags:
      - 管理端｜管理员认证
    put:
      consumes:
      - application/json
      description: 当前登录管理员更新自己的昵称、邮箱、头像
      parameters:
      - description: 更新资料请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/admin.Profile'
              type: object
      security:
      - BearerAuth: []
      summary: 更新管理员个人资料
      tags:
      - 管理端｜管理员认证
  /api/admin/v1/auth/profile/password:
    put:
      consumes:
      - application/json
      description: 当前登录管理员修改自己的密码
      parameters:
      - description: 修改密码请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 修改管理员密码
      tags:
      - 管理端｜管理员认证
...
  /api/admin/v1/categories:
    get:
      description: 获取分类树形列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/admin.CategoryResponse'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 分类树
      tags:
      - 管理端｜分类管理
    post:
      consumes:
      - application/json
      description: 创建一个新的分类
      parameters:
      - description: 分类参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/admin.CategoryResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 新建分类
      tags:
      - 管理端｜分类管理
  /api/admin/v1/categories/{id}:
    delete:
      description: 删除指定分类
      parameters:
      - description: 分类ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除分类
      tags:
      - 管理端｜分类管理
    put:
      consumes:
      - application/json
      description: 更新指定分类信息
      parameters:
      - description: 分类ID
        in: path
        name: id
        required: true
        type: integer
      - description: 分类参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/admin.CategoryResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 更新分类
      tags:
      - 管理端｜分类管理
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"